refactor: nested classmap classes into FileParser class
