+++ src/ServiceLayer.Mesh/FileTypes/NbssAppointmentEvents/FileParser.cs
@@ -108,4 +108,27 @@ public class FileParser : IFileParser
 
         return record;
     }
-}
+
+    public sealed class FileTrailerRecordMap : ClassMap<FileTrailerRecord>
+    {
+        public FileTrailerRecordMap()
+        {
+            Map(m => m.RecordTypeIdentifier).Index(0);
+            Map(m => m.ExtractId).Index(1);
+            Map(m => m.TransferEndDate).Index(2);
+            Map(m => m.TransferEndTime).Index(3);
+            Map(m => m.RecordCount).Index(4);
+        }
+    }
+
+    public sealed class FileHeaderRecordMap : ClassMap<FileHeaderRecord>
+    {
+        public FileHeaderRecordMap()
+        {
+            Map(m => m.RecordTypeIdentifier).Index(0);
+            Map(m => m.ExtractId).Index(1);
+            Map(m => m.TransferStartDate).Index(2);
+            Map(m => m.TransferStartTime).Index(3);
+            Map(m => m.RecordCount).Index(4);
+        }
+    }
